Make domain participant fields private
