+++ dds/src/implementation/domain_participant_backend/entities/mod.rs
@@ -1,5 +1,6 @@
 pub mod data_reader;
 pub mod data_writer;
+pub mod domain_participant;
 pub mod publisher;
 pub mod subscriber;
 pub mod topic;
