+++ dds/src/implementation/domain_participant_backend/services/domain_participant_service.rs
@@ -2,9 +2,8 @@ use std::sync::Arc;
 
 use crate::{
     builtin_topics::{
-        BuiltInTopicKey, ParticipantBuiltinTopicData, PublicationBuiltinTopicData,
-        SubscriptionBuiltinTopicData, TopicBuiltinTopicData, DCPS_PARTICIPANT, DCPS_PUBLICATION,
-        DCPS_SUBSCRIPTION, DCPS_TOPIC,
+        BuiltInTopicKey, ParticipantBuiltinTopicData, TopicBuiltinTopicData, DCPS_PARTICIPANT,
+        DCPS_PUBLICATION, DCPS_SUBSCRIPTION, DCPS_TOPIC,
     },
     dds_async::{
         domain_participant_listener::DomainParticipantListenerAsync,
@@ -14,7 +13,7 @@ use crate::{
     implementation::{
         actor::{Actor, ActorAddress, Mail, MailHandler},
         domain_participant_backend::{
-            domain_participant_actor::{DomainParticipantActor, ParticipantListenerThread},
+            domain_participant_actor::DomainParticipantActor,
             entities::{publisher::PublisherActor, subscriber::SubscriberActor, topic::TopicActor},
         },
         listeners::{
@@ -54,7 +53,7 @@ impl MailHandler<CreateUserDefinedPublisher> for DomainParticipantActor {
         message: CreateUserDefinedPublisher,
     ) -> <CreateUserDefinedPublisher as Mail>::Result {
         let publisher_qos = match message.qos {
-            QosKind::Default => self.default_publisher_qos.clone(),
+            QosKind::Default => self.domain_participant.default_publisher_qos().clone(),
             QosKind::Specific(q) => q,
         };
 
@@ -70,11 +69,17 @@ impl MailHandler<CreateUserDefinedPublisher> for DomainParticipantActor {
             status_condition,
         );
 
-        if self.enabled && self.qos.entity_factory.autoenable_created_entities {
+        if self.domain_participant.enabled()
+            && self
+                .domain_participant
+                .qos()
+                .entity_factory
+                .autoenable_created_entities
+        {
             publisher.enable();
         }
 
-        self.user_defined_publisher_list.push(publisher);
+        self.domain_participant.insert_publisher(publisher);
 
         Ok((publisher_handle, publisher_status_condition_address))
     }
@@ -92,32 +97,29 @@ impl MailHandler<DeleteUserDefinedPublisher> for DomainParticipantActor {
         &mut self,
         message: DeleteUserDefinedPublisher,
     ) -> <DeleteUserDefinedPublisher as Mail>::Result {
-        if message.participant_handle != self.get_instance_handle() {
+        if message.participant_handle != self.domain_participant.get_instance_handle() {
             return Err(DdsError::PreconditionNotMet(
                 "Publisher can only be deleted from its parent participant".to_string(),
             ));
         }
 
-        if let Some(i) = self
-            .user_defined_publisher_list
-            .iter()
-            .position(|p| p.instance_handle() == message.publisher_handle)
+        if self
+            .domain_participant
+            .get_publisher(message.publisher_handle)
+            .ok_or(DdsError::AlreadyDeleted)?
+            .data_writer_list()
+            .count()
+            > 0
         {
-            if self.user_defined_publisher_list[i]
-                .data_writer_list()
-                .count()
-                == 0
-            {
-                self.user_defined_publisher_list.remove(i);
-                Ok(())
-            } else {
-                Err(DdsError::PreconditionNotMet(
-                    "Publisher still contains data writers".to_string(),
-                ))
-            }
-        } else {
-            Err(DdsError::AlreadyDeleted)
+            return Err(DdsError::PreconditionNotMet(
+                "Publisher still contains data writers".to_string(),
+            ));
         }
+        self.domain_participant
+            .remove_publisher(&message.publisher_handle)
+            .ok_or(DdsError::AlreadyDeleted)?;
+
+        Ok(())
     }
 }
 
@@ -135,7 +137,7 @@ impl MailHandler<CreateUserDefinedSubscriber> for DomainParticipantActor {
         message: CreateUserDefinedSubscriber,
     ) -> <CreateUserDefinedSubscriber as Mail>::Result {
         let subscriber_qos = match message.qos {
-            QosKind::Default => self.default_subscriber_qos.clone(),
+            QosKind::Default => self.domain_participant.default_subscriber_qos().clone(),
             QosKind::Specific(q) => q,
         };
         let subscriber_handle = self.instance_handle_counter.generate_new_instance_handle();
@@ -151,11 +153,17 @@ impl MailHandler<CreateUserDefinedSubscriber> for DomainParticipantActor {
 
         let subscriber_status_condition_address = subscriber.status_condition().address();
 
-        if self.enabled && self.qos.entity_factory.autoenable_created_entities {
+        if self.domain_participant.enabled()
+            && self
+                .domain_participant
+                .qos()
+                .entity_factory
+                .autoenable_created_entities
+        {
             subscriber.enable();
         }
 
-        self.user_defined_subscriber_list.push(subscriber);
+        self.domain_participant.insert_subscriber(subscriber);
 
         Ok((
             subscriber_handle.into(),
@@ -176,32 +184,29 @@ impl MailHandler<DeleteUserDefinedSubscriber> for DomainParticipantActor {
         &mut self,
         message: DeleteUserDefinedSubscriber,
     ) -> <DeleteUserDefinedSubscriber as Mail>::Result {
-        if self.get_instance_handle() != message.participant_handle {
+        if self.domain_participant.get_instance_handle() != message.participant_handle {
             return Err(DdsError::PreconditionNotMet(
                 "Subscriber can only be deleted from its parent participant".to_string(),
             ));
         }
 
-        if let Some(i) = self
-            .user_defined_subscriber_list
-            .iter()
-            .position(|s| s.instance_handle() == message.subscriber_handle)
+        if self
+            .domain_participant
+            .get_subscriber(message.subscriber_handle)
+            .ok_or(DdsError::AlreadyDeleted)?
+            .data_reader_list()
+            .count()
+            > 0
         {
-            if self.user_defined_subscriber_list[i]
-                .data_reader_list()
-                .count()
-                == 0
-            {
-                self.user_defined_subscriber_list.remove(i);
-                Ok(())
-            } else {
-                Err(DdsError::PreconditionNotMet(
-                    "Subscriber still contains data readers".to_string(),
-                ))
-            }
-        } else {
-            Err(DdsError::AlreadyDeleted)
+            return Err(DdsError::PreconditionNotMet(
+                "Subscriber still contains data readers".to_string(),
+            ));
         }
+        self.domain_participant
+            .remove_subscriber(&message.subscriber_handle)
+            .ok_or(DdsError::AlreadyDeleted)?;
+
+        Ok(())
     }
 }
 
@@ -221,7 +226,11 @@ impl MailHandler<CreateUserDefinedTopic> for DomainParticipantActor {
         &mut self,
         message: CreateUserDefinedTopic,
     ) -> <CreateUserDefinedTopic as Mail>::Result {
-        if self.topic_list.contains_key(&message.topic_name) {
+        if self
+            .domain_participant
+            .get_topic(&message.topic_name)
+            .is_some()
+        {
             return Err(DdsError::PreconditionNotMet(format!(
                 "Topic with name {} already exists.
              To access this topic call the lookup_topicdescription method.",
@@ -230,7 +239,7 @@ impl MailHandler<CreateUserDefinedTopic> for DomainParticipantActor {
         }
 
         let qos = match message.qos {
-            QosKind::Default => self.default_topic_qos.clone(),
+            QosKind::Default => self.domain_participant.get_default_topic_qos().clone(),
             QosKind::Specific(q) => q,
         };
 
@@ -250,7 +259,13 @@ impl MailHandler<CreateUserDefinedTopic> for DomainParticipantActor {
             message.type_support,
         );
 
-        if self.enabled && self.qos.entity_factory.autoenable_created_entities {
+        if self.domain_participant.enabled()
+            && self
+                .domain_participant
+                .qos()
+                .entity_factory
+                .autoenable_created_entities
+        {
             let topic_qos = topic.qos();
             let topic_builtin_topic_data = TopicBuiltinTopicData {
                 key: BuiltInTopicKey {
@@ -272,10 +287,11 @@ impl MailHandler<CreateUserDefinedTopic> for DomainParticipantActor {
                 topic_data: topic_qos.topic_data.clone(),
                 representation: topic_qos.representation.clone(),
             };
-            self.announce_topic(topic_builtin_topic_data)?;
+            self.domain_participant
+                .announce_topic(topic_builtin_topic_data)?;
         }
 
-        self.topic_list.insert(message.topic_name, topic);
+        self.domain_participant.insert_topic(topic);
 
         Ok((topic_handle, topic_status_condition_address))
     }
@@ -293,7 +309,7 @@ impl MailHandler<DeleteUserDefinedTopic> for DomainParticipantActor {
         &mut self,
         message: DeleteUserDefinedTopic,
     ) -> <DeleteUserDefinedTopic as Mail>::Result {
-        if self.get_instance_handle() != message.participant_handle {
+        if self.domain_participant.get_instance_handle() != message.participant_handle {
             return Err(DdsError::PreconditionNotMet(
                 "Topic can only be deleted from its parent participant".to_string(),
             ));
@@ -303,32 +319,20 @@ impl MailHandler<DeleteUserDefinedTopic> for DomainParticipantActor {
             return Ok(());
         }
 
-        for publisher in self.user_defined_publisher_list.iter_mut() {
-            if publisher
-                .data_writer_list_mut()
-                .find(|dw| dw.topic_name() == message.topic_name)
-                .is_some()
-            {
-                return Err(DdsError::PreconditionNotMet(
-                    "Topic still attached to some data writer".to_string(),
-                ));
-            }
+        if Arc::strong_count(
+            self.domain_participant
+                .get_topic(&message.topic_name)
+                .ok_or(DdsError::AlreadyDeleted)?
+                .type_support(),
+        ) > 1
+        {
+            return Err(DdsError::PreconditionNotMet(
+                "Topic still attached to some data writer or data reader".to_string(),
+            ));
         }
 
-        for subscriber in self.user_defined_subscriber_list.iter_mut() {
-            if subscriber
-                .data_reader_list()
-                .find(|dr| dr.topic_name() == message.topic_name)
-                .is_some()
-            {
-                return Err(DdsError::PreconditionNotMet(
-                    "Topic still attached to some data reader".to_string(),
-                ));
-            }
-        }
-
-        self.topic_list
-            .remove(&message.topic_name)
+        self.domain_participant
+            .remove_topic(&message.topic_name)
             .ok_or(DdsError::AlreadyDeleted)?;
 
         Ok(())
@@ -344,52 +348,52 @@ impl Mail for FindTopic {
 }
 impl MailHandler<FindTopic> for DomainParticipantActor {
     fn handle(&mut self, message: FindTopic) -> <FindTopic as Mail>::Result {
-        if let Some(topic) = self.topic_list.get(&message.topic_name) {
+        if let Some(topic) = self.domain_participant.get_topic(&message.topic_name) {
             Ok(Some((
                 topic.instance_handle().into(),
                 topic.status_condition().address(),
                 topic.type_name().to_owned(),
             )))
         } else {
-            for discovered_topic_data in self.discovered_topic_list.values() {
-                if discovered_topic_data.name() == message.topic_name {
-                    let qos = TopicQos {
-                        topic_data: discovered_topic_data.topic_data().clone(),
-                        durability: discovered_topic_data.durability().clone(),
-                        deadline: discovered_topic_data.deadline().clone(),
-                        latency_budget: discovered_topic_data.latency_budget().clone(),
-                        liveliness: discovered_topic_data.liveliness().clone(),
-                        reliability: discovered_topic_data.reliability().clone(),
-                        destination_order: discovered_topic_data.destination_order().clone(),
-                        history: discovered_topic_data.history().clone(),
-                        resource_limits: discovered_topic_data.resource_limits().clone(),
-                        transport_priority: discovered_topic_data.transport_priority().clone(),
-                        lifespan: discovered_topic_data.lifespan().clone(),
-                        ownership: discovered_topic_data.ownership().clone(),
-                        representation: discovered_topic_data.representation().clone(),
-                    };
-                    let type_name = discovered_topic_data.type_name.clone();
-                    let topic_handle = self.instance_handle_counter.generate_new_instance_handle();
-                    let mut topic = TopicActor::new(
-                        qos,
-                        type_name.clone(),
-                        message.topic_name.clone(),
-                        topic_handle,
-                        Actor::spawn(StatusConditionActor::default(), &self.executor.handle()),
-                        None,
-                        vec![],
-                        message.type_support,
-                    );
-                    topic.enable();
-                    let topic_status_condition_address = topic.status_condition().address();
+            if let Some(discovered_topic_data) =
+                self.domain_participant.find_topic(&message.topic_name)
+            {
+                let qos = TopicQos {
+                    topic_data: discovered_topic_data.topic_data().clone(),
+                    durability: discovered_topic_data.durability().clone(),
+                    deadline: discovered_topic_data.deadline().clone(),
+                    latency_budget: discovered_topic_data.latency_budget().clone(),
+                    liveliness: discovered_topic_data.liveliness().clone(),
+                    reliability: discovered_topic_data.reliability().clone(),
+                    destination_order: discovered_topic_data.destination_order().clone(),
+                    history: discovered_topic_data.history().clone(),
+                    resource_limits: discovered_topic_data.resource_limits().clone(),
+                    transport_priority: discovered_topic_data.transport_priority().clone(),
+                    lifespan: discovered_topic_data.lifespan().clone(),
+                    ownership: discovered_topic_data.ownership().clone(),
+                    representation: discovered_topic_data.representation().clone(),
+                };
+                let type_name = discovered_topic_data.type_name.clone();
+                let topic_handle = self.instance_handle_counter.generate_new_instance_handle();
+                let mut topic = TopicActor::new(
+                    qos,
+                    type_name.clone(),
+                    message.topic_name.clone(),
+                    topic_handle,
+                    Actor::spawn(StatusConditionActor::default(), &self.executor.handle()),
+                    None,
+                    vec![],
+                    message.type_support,
+                );
+                topic.enable();
+                let topic_status_condition_address = topic.status_condition().address();
 
-                    self.topic_list.insert(message.topic_name, topic);
-                    return Ok(Some((
-                        topic_handle.into(),
-                        topic_status_condition_address,
-                        type_name,
-                    )));
-                }
+                self.domain_participant.insert_topic(topic);
+                return Ok(Some((
+                    topic_handle.into(),
+                    topic_status_condition_address,
+                    type_name,
+                )));
             }
             Ok(None)
         }
@@ -407,7 +411,7 @@ impl MailHandler<LookupTopicdescription> for DomainParticipantActor {
         &mut self,
         message: LookupTopicdescription,
     ) -> <LookupTopicdescription as Mail>::Result {
-        if let Some(topic) = self.topic_list.get(&message.topic_name) {
+        if let Some(topic) = self.domain_participant.get_topic(&message.topic_name) {
             Ok(Some((
                 topic.type_name().to_owned(),
                 topic.instance_handle().into(),
@@ -427,8 +431,8 @@ impl Mail for IgnoreParticipant {
 }
 impl MailHandler<IgnoreParticipant> for DomainParticipantActor {
     fn handle(&mut self, message: IgnoreParticipant) -> <IgnoreParticipant as Mail>::Result {
-        if self.enabled {
-            self.ignored_participants.insert(message.handle);
+        if self.domain_participant.enabled() {
+            self.domain_participant.ignore_participant(message.handle);
             Ok(())
         } else {
             Err(DdsError::NotEnabled)
@@ -444,8 +448,8 @@ impl Mail for IgnoreSubscription {
 }
 impl MailHandler<IgnoreSubscription> for DomainParticipantActor {
     fn handle(&mut self, message: IgnoreSubscription) -> <IgnoreSubscription as Mail>::Result {
-        if self.enabled {
-            self.ignored_subcriptions.insert(message.handle);
+        if self.domain_participant.enabled() {
+            self.domain_participant.ignore_subscription(message.handle);
             Ok(())
         } else {
             Err(DdsError::NotEnabled)
@@ -461,8 +465,8 @@ impl Mail for IgnorePublication {
 }
 impl MailHandler<IgnorePublication> for DomainParticipantActor {
     fn handle(&mut self, message: IgnorePublication) -> <IgnorePublication as Mail>::Result {
-        if self.enabled {
-            self.ignored_publications.insert(message.handle);
+        if self.domain_participant.enabled() {
+            self.domain_participant.ignore_publication(message.handle);
             Ok(())
         } else {
             Err(DdsError::NotEnabled)
@@ -480,77 +484,78 @@ impl MailHandler<DeleteParticipantContainedEntities> for DomainParticipantActor 
         _: DeleteParticipantContainedEntities,
     ) -> <DeleteParticipantContainedEntities as Mail>::Result {
         let deleted_publisher_list: Vec<PublisherActor> =
-            self.user_defined_publisher_list.drain(..).collect();
+            self.domain_participant.drain_publisher_list().collect();
         for mut publisher in deleted_publisher_list {
             let publisher_qos = publisher.qos().clone();
-            for data_writer in publisher.drain_data_writer_list() {
-                let publication_builtin_topic_data = PublicationBuiltinTopicData {
-                    key: BuiltInTopicKey {
-                        value: data_writer.transport_writer().guid(),
-                    },
-                    participant_key: BuiltInTopicKey { value: [0; 16] },
-                    topic_name: data_writer.topic_name().to_owned(),
-                    type_name: data_writer.type_name().to_owned(),
-                    durability: data_writer.qos().durability.clone(),
-                    deadline: data_writer.qos().deadline.clone(),
-                    latency_budget: data_writer.qos().latency_budget.clone(),
-                    liveliness: data_writer.qos().liveliness.clone(),
-                    reliability: data_writer.qos().reliability.clone(),
-                    lifespan: data_writer.qos().lifespan.clone(),
-                    user_data: data_writer.qos().user_data.clone(),
-                    ownership: data_writer.qos().ownership.clone(),
-                    ownership_strength: data_writer.qos().ownership_strength.clone(),
-                    destination_order: data_writer.qos().destination_order.clone(),
-                    presentation: publisher_qos.presentation.clone(),
-                    partition: publisher_qos.partition.clone(),
-                    topic_data: self.topic_list[data_writer.topic_name()]
-                        .qos()
-                        .topic_data
-                        .clone(),
-                    group_data: publisher_qos.group_data.clone(),
-                    representation: data_writer.qos().representation.clone(),
-                };
-                self.announce_deleted_data_writer(publication_builtin_topic_data)?;
-            }
+            todo!();
+            // for data_writer in publisher.drain_data_writer_list() {
+            //     let publication_builtin_topic_data = PublicationBuiltinTopicData {
+            //         key: BuiltInTopicKey {
+            //             value: data_writer.transport_writer().guid(),
+            //         },
+            //         participant_key: BuiltInTopicKey { value: [0; 16] },
+            //         topic_name: data_writer.topic_name().to_owned(),
+            //         type_name: data_writer.type_name().to_owned(),
+            //         durability: data_writer.qos().durability.clone(),
+            //         deadline: data_writer.qos().deadline.clone(),
+            //         latency_budget: data_writer.qos().latency_budget.clone(),
+            //         liveliness: data_writer.qos().liveliness.clone(),
+            //         reliability: data_writer.qos().reliability.clone(),
+            //         lifespan: data_writer.qos().lifespan.clone(),
+            //         user_data: data_writer.qos().user_data.clone(),
+            //         ownership: data_writer.qos().ownership.clone(),
+            //         ownership_strength: data_writer.qos().ownership_strength.clone(),
+            //         destination_order: data_writer.qos().destination_order.clone(),
+            //         presentation: publisher_qos.presentation.clone(),
+            //         partition: publisher_qos.partition.clone(),
+            //         topic_data: self.topic_list[data_writer.topic_name()]
+            //             .qos()
+            //             .topic_data
+            //             .clone(),
+            //         group_data: publisher_qos.group_data.clone(),
+            //         representation: data_writer.qos().representation.clone(),
+            //     };
+            //     self.announce_deleted_data_writer(publication_builtin_topic_data)?;
+            // }
         }
 
         let deleted_subscriber_list: Vec<SubscriberActor> =
-            self.user_defined_subscriber_list.drain(..).collect();
+            self.domain_participant.drain_subscriber_list().collect();
         for mut subscriber in deleted_subscriber_list {
             let subscriber_qos = subscriber.qos().clone();
             for data_reader in subscriber.drain_data_reader_list() {
-                let subscription_builtin_topic_data = SubscriptionBuiltinTopicData {
-                    key: BuiltInTopicKey {
-                        value: data_reader.transport_reader().guid(),
-                    },
-                    participant_key: BuiltInTopicKey { value: [0; 16] },
-                    topic_name: data_reader.topic_name().to_owned(),
-                    type_name: data_reader.type_name().to_owned(),
-                    durability: data_reader.qos().durability.clone(),
-                    deadline: data_reader.qos().deadline.clone(),
-                    latency_budget: data_reader.qos().latency_budget.clone(),
-                    liveliness: data_reader.qos().liveliness.clone(),
-                    reliability: data_reader.qos().reliability.clone(),
-                    ownership: data_reader.qos().ownership.clone(),
-                    destination_order: data_reader.qos().destination_order.clone(),
-                    user_data: data_reader.qos().user_data.clone(),
-                    time_based_filter: data_reader.qos().time_based_filter.clone(),
-                    presentation: subscriber_qos.presentation.clone(),
-                    partition: subscriber_qos.partition.clone(),
-                    topic_data: self.topic_list[data_reader.topic_name()]
-                        .qos()
-                        .topic_data
-                        .clone(),
-                    group_data: subscriber_qos.group_data.clone(),
-                    representation: data_reader.qos().representation.clone(),
-                };
+                todo!()
+                // let subscription_builtin_topic_data = SubscriptionBuiltinTopicData {
+                //     key: BuiltInTopicKey {
+                //         value: data_reader.transport_reader().guid(),
+                //     },
+                //     participant_key: BuiltInTopicKey { value: [0; 16] },
+                //     topic_name: data_reader.topic_name().to_owned(),
+                //     type_name: data_reader.type_name().to_owned(),
+                //     durability: data_reader.qos().durability.clone(),
+                //     deadline: data_reader.qos().deadline.clone(),
+                //     latency_budget: data_reader.qos().latency_budget.clone(),
+                //     liveliness: data_reader.qos().liveliness.clone(),
+                //     reliability: data_reader.qos().reliability.clone(),
+                //     ownership: data_reader.qos().ownership.clone(),
+                //     destination_order: data_reader.qos().destination_order.clone(),
+                //     user_data: data_reader.qos().user_data.clone(),
+                //     time_based_filter: data_reader.qos().time_based_filter.clone(),
+                //     presentation: subscriber_qos.presentation.clone(),
+                //     partition: subscriber_qos.partition.clone(),
+                //     topic_data: self.topic_list[data_reader.topic_name()]
+                //         .qos()
+                //         .topic_data
+                //         .clone(),
+                //     group_data: subscriber_qos.group_data.clone(),
+                //     representation: data_reader.qos().representation.clone(),
+                // };
 
-                self.announce_deleted_data_reader(subscription_builtin_topic_data)?;
+                // self.announce_deleted_data_reader(subscription_builtin_topic_data)?;
             }
         }
 
-        self.topic_list
-            .retain(|_, x| BUILT_IN_TOPIC_NAME_LIST.contains(&x.topic_name()));
+        self.domain_participant.delete_all_topics();
 
         Ok(())
     }
@@ -572,7 +577,7 @@ impl MailHandler<SetDefaultPublisherQos> for DomainParticipantActor {
             QosKind::Specific(q) => q,
         };
 
-        self.default_publisher_qos = qos;
+        self.domain_participant.set_default_publisher_qos(qos);
         Ok(())
     }
 }
@@ -583,7 +588,7 @@ impl Mail for GetDefaultPublisherQos {
 }
 impl MailHandler<GetDefaultPublisherQos> for DomainParticipantActor {
     fn handle(&mut self, _: GetDefaultPublisherQos) -> <GetDefaultPublisherQos as Mail>::Result {
-        Ok(self.default_publisher_qos.clone())
+        Ok(self.domain_participant.default_publisher_qos().clone())
     }
 }
 
@@ -603,7 +608,7 @@ impl MailHandler<SetDefaultSubscriberQos> for DomainParticipantActor {
             QosKind::Specific(q) => q,
         };
 
-        self.default_subscriber_qos = qos;
+        self.domain_participant.set_default_subscriber_qos(qos);
 
         Ok(())
     }
@@ -615,7 +620,7 @@ impl Mail for GetDefaultSubscriberQos {
 }
 impl MailHandler<GetDefaultSubscriberQos> for DomainParticipantActor {
     fn handle(&mut self, _: GetDefaultSubscriberQos) -> <GetDefaultSubscriberQos as Mail>::Result {
-        Ok(self.default_subscriber_qos.clone())
+        Ok(self.domain_participant.default_subscriber_qos().clone())
     }
 }
 
@@ -635,7 +640,7 @@ impl MailHandler<SetDefaultTopicQos> for DomainParticipantActor {
             }
         };
 
-        self.default_topic_qos = qos;
+        self.domain_participant.set_default_topic_qos(qos);
 
         Ok(())
     }
@@ -647,7 +652,7 @@ impl Mail for GetDefaultTopicQos {
 }
 impl MailHandler<GetDefaultTopicQos> for DomainParticipantActor {
     fn handle(&mut self, _: GetDefaultTopicQos) -> <GetDefaultTopicQos as Mail>::Result {
-        Ok(self.default_topic_qos.clone())
+        Ok(self.domain_participant.get_default_topic_qos().clone())
     }
 }
 
@@ -660,7 +665,7 @@ impl MailHandler<GetDiscoveredParticipants> for DomainParticipantActor {
         &mut self,
         _: GetDiscoveredParticipants,
     ) -> <GetDiscoveredParticipants as Mail>::Result {
-        Ok(self.discovered_participant_list.keys().cloned().collect())
+        Ok(self.domain_participant.get_discovered_participants())
     }
 }
 
@@ -676,8 +681,8 @@ impl MailHandler<GetDiscoveredParticipantData> for DomainParticipantActor {
         message: GetDiscoveredParticipantData,
     ) -> <GetDiscoveredParticipantData as Mail>::Result {
         Ok(self
-            .discovered_participant_list
-            .get(&message.participant_handle)
+            .domain_participant
+            .get_discovered_participant_data(&message.participant_handle)
             .ok_or(DdsError::BadParameter)?
             .dds_participant_data
             .clone())
@@ -690,7 +695,7 @@ impl Mail for GetDiscoveredTopics {
 }
 impl MailHandler<GetDiscoveredTopics> for DomainParticipantActor {
     fn handle(&mut self, _: GetDiscoveredTopics) -> <GetDiscoveredTopics as Mail>::Result {
-        Ok(self.discovered_topic_list.keys().cloned().collect())
+        Ok(self.domain_participant.get_discovered_topics())
     }
 }
 
@@ -705,8 +710,8 @@ impl MailHandler<GetDiscoveredTopicData> for DomainParticipantActor {
         &mut self,
         message: GetDiscoveredTopicData,
     ) -> <GetDiscoveredTopicData as Mail>::Result {
-        self.discovered_topic_list
-            .get(&message.topic_handle)
+        self.domain_participant
+            .get_discovered_topic_data(&message.topic_handle)
             .cloned()
             .ok_or(DdsError::PreconditionNotMet(
                 "Topic with this handle not discovered".to_owned(),
@@ -720,7 +725,7 @@ impl Mail for GetCurrentTime {
 }
 impl MailHandler<GetCurrentTime> for DomainParticipantActor {
     fn handle(&mut self, _: GetCurrentTime) -> <GetCurrentTime as Mail>::Result {
-        self.get_current_time()
+        self.domain_participant.get_current_time()
     }
 }
 
@@ -740,8 +745,8 @@ impl MailHandler<SetDomainParticipantQos> for DomainParticipantActor {
             QosKind::Specific(q) => q,
         };
 
-        self.qos = qos;
-        self.announce_participant()
+        self.domain_participant.set_qos(qos);
+        self.domain_participant.announce_participant()
     }
 }
 
@@ -751,7 +756,7 @@ impl Mail for GetDomainParticipantQos {
 }
 impl MailHandler<GetDomainParticipantQos> for DomainParticipantActor {
     fn handle(&mut self, _: GetDomainParticipantQos) -> <GetDomainParticipantQos as Mail>::Result {
-        Ok(self.qos.clone())
+        Ok(self.domain_participant.qos().clone())
     }
 }
 
@@ -767,12 +772,13 @@ impl MailHandler<SetDomainParticipantListener> for DomainParticipantActor {
         &mut self,
         message: SetDomainParticipantListener,
     ) -> <SetDomainParticipantListener as Mail>::Result {
-        if let Some(l) = self.participant_listener_thread.take() {
-            l.join()?;
-        }
-        self.participant_listener_thread = message.listener.map(ParticipantListenerThread::new);
-        self.status_kind = message.status_kind;
-        Ok(())
+        todo!()
+        // if let Some(l) = self.participant_listener_thread.take() {
+        //     l.join()?;
+        // }
+        // self.participant_listener_thread = message.listener.map(ParticipantListenerThread::new);
+        // self.status_kind = message.status_kind;
+        // Ok(())
     }
 }
 
@@ -782,9 +788,9 @@ impl Mail for EnableDomainParticipant {
 }
 impl MailHandler<EnableDomainParticipant> for DomainParticipantActor {
     fn handle(&mut self, _: EnableDomainParticipant) -> <EnableDomainParticipant as Mail>::Result {
-        if !self.enabled {
-            self.enabled = true;
-            self.announce_participant()?;
+        if !self.domain_participant.enabled() {
+            self.domain_participant.enable();
+            self.domain_participant.announce_participant()?;
         }
         Ok(())
     }
@@ -796,15 +802,6 @@ impl Mail for IsEmpty {
 }
 impl MailHandler<IsEmpty> for DomainParticipantActor {
     fn handle(&mut self, _: IsEmpty) -> <IsEmpty as Mail>::Result {
-        let no_user_defined_topics = self
-            .topic_list
-            .keys()
-            .filter(|t| !BUILT_IN_TOPIC_NAME_LIST.contains(&t.as_ref()))
-            .count()
-            == 0;
-
-        self.user_defined_publisher_list.is_empty()
-            && self.user_defined_subscriber_list.is_empty()
-            && no_user_defined_topics
+        self.domain_participant.is_empty()
     }
 }
