+++ dds/src/implementation/domain_participant_backend/services/mod.rs
@@ -2,6 +2,7 @@ pub mod data_reader_service;
 pub mod data_writer_service;
 pub mod domain_participant_service;
 pub mod event_service;
+pub mod message_service;
 pub mod publisher_service;
 pub mod subscriber_service;
 pub mod topic_service;
